refactor: benchmark kinds have own collector func
