+++ packages/graphqlgen/benchmarks/integration/index.ts
@@ -10,6 +10,45 @@ import * as GGen from '../../src'
 import * as Bench from 'benchmark'
 import * as Path from 'path'
 import * as Glob from 'glob'
+import * as Util from '../../src/utils'
+
+const collect = (): Bench[] => {
+  const paths = Glob.sync(Path.join(__dirname, './*'))
+  const benchmarks: Bench[] = []
+
+  for (const path of paths) {
+    if (Util.isFile(path)) continue
+
+    const errors = validateFixtures(path)
+
+    if (errors) {
+      for (const error of errors) {
+        console.log(error.message)
+        process.exit(1)
+      }
+    }
+
+    const modelPaths = Glob.sync(Path.join(path, './models.*'))
+
+    for (const modelPath of modelPaths) {
+      // 1. We know there will be an extension because of isFile
+      // filter above. 2. We know it will be a support language extension
+      // because of the filter right after.
+      const ext = Util.getExt(modelPath) as Util.LanguageExtension
+
+      if (!Util.languageExtensions.includes(ext)) continue
+
+      const benchmark = createBenchmark({
+        language: Util.getLangFromExt(ext),
+        rootPath: path,
+      })
+
+      benchmarks.push(benchmark)
+    }
+  }
+
+  return benchmarks
+}
 
 type Options = {
   language: ConfigTypes.GraphQLGenDefinition['language']
@@ -83,21 +122,28 @@ const createCodeGenConfig = (
   }
 }
 
-const validateFixtures = (scenarioFolder: string): void => {
+/**
+ * Function that checks for correctness of folder/file layout of benchmarks.
+ * The returned validation failure, if any, contains a message to help developers
+ * fix the problem and so should be printed to them nicely.
+ */
+const validateFixtures = (scenarioFolder: string): null | Error[] => {
   const name = Path.basename(scenarioFolder)
   const files = Glob.sync(Path.join(scenarioFolder, './*')).map(path =>
     Path.basename(path),
   )
 
-  // TODO throw a multi-error
+  const errors = []
 
   if (!files.includes('schema.graphql')) {
-    throw new Error(`benchmark "${name}" missing file schema.graphql`)
+    errors.push(new Error(`benchmark "${name}" missing file schema.graphql`))
   }
 
   if (!files.includes('models.ts')) {
-    throw new Error(`benchmark "${name}" missing file models.ts`)
+    errors.push(new Error(`benchmark "${name}" missing file models.ts`))
   }
+
+  return errors.length ? errors : null
 }
 
-export { validateFixtures, createCodeGenConfig as setup, createBenchmark }
+export { collect }
